proyecto_vision_modular_final/requirements.txt
numpy
opencv-python
tensorflow
pandas
Pillow
openpyxl
matplotlib
pytest
google-cloud-vision
gspread
google-auth


